+++ finagle-http/src/test/java/com/twitter/finagle/http2/param/StackParamCompilationTest.java
@@ -6,12 +6,15 @@ import org.junit.Test;
 
 import com.twitter.finagle.Http;
 
+import static com.twitter.util.Function.func;
+
 public class StackParamCompilationTest {
 
   @Test
   public void testParams() {
     Http.client()
       .withStack(Http.client().stack())
+      .withStack(func(stack -> stack))
       .configured(new MaxFrameSize(Option.empty()).mk());
   }